P38 Exception

+++ src/main/java/life/majiang/community/controller/PublishController.java
@@ -34,7 +34,7 @@ public class PublishController {
 //    private UserMapper userMapper;
 
     @GetMapping("/publish/{id}")
-    public String eit(@PathVariable(name = "id") Integer id,
+    public String eit(@PathVariable(name = "id") Long id,
                       Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
@@ -53,7 +53,7 @@ public class PublishController {
     public String doPublish(@RequestParam(value = "title", required = false) String title,
                             @RequestParam(value = "description",required = false) String description,
                             @RequestParam(value = "tag",required = false) String tag,
-                            @RequestParam(value = "id",required = false) Integer id,
+                            @RequestParam(value = "id",required = false) Long id,
                             HttpServletRequest request,
                             Model model){
