+++ src/main/java/life/majiang/community/dto/QuestionDTO.java
@@ -9,12 +9,14 @@ import lombok.Data;
  */
 @Data
 public class QuestionDTO {
-    private Integer id;
+//    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creator;
+//    private Integer creator;
+    private Long creator;
     private Integer commentCount;
     private Integer viewCount;
     private Integer likeCount;
